tests/test_chain.py
from chain import get_openai_llm, get_huggingface_hub_llm, get_google_llm, get_chain
from unittest.mock import patch, MagicMock
from langchain.chains import LLMChain

# Mocking external classes and functions
@patch('chain.ChatOpenAI')
def test_get_openai_llm(mock_chat_openai):
    get_openai_llm(model="gpt-3.5-turbo")
    mock_chat_openai.assert_called_with(model="gpt-3.5-turbo", temperature=0)

@patch('chain.HuggingFaceHub')
def test_get_huggingface_hub_llm(mock_huggingface_hub):
    get_huggingface_hub_llm(repo_id="mistralai/Mixtral-8x7B-Instruct-v0.1")
    mock_huggingface_hub.assert_called_with(repo_id="mistralai/Mixtral-8x7B-Instruct-v0.1", model_kwargs={"temperature":0.2, "max_length":1000})

@patch('chain.ChatGoogleGenerativeAI')
def test_get_google_llm(mock_chat_google_genai):
    get_google_llm(model="gemini-pro")
    # Here, add assertions depending on the implementation of get_google_llm

"""
@patch('chain.ChatOpenAI')
@patch('chain.HuggingFaceHub')
@patch('chain.ChatGoogleGenerativeAI')
def test_get_chain(mock_chat_google_genai, mock_huggingface_hub, mock_chat_openai):
    chain = get_chain(model="gpt-3.5-turbo", chain_type="stuff")
    assert isinstance(chain, LLMChain)
    mock_chat_openai.assert_called_with(model="gpt-3.5-turbo", temperature=0)

    chain = get_chain(model="mistralai/Mixtral-8x7B-Instruct-v0.1", chain_type="stuff")
    mock_huggingface_hub.assert_called_with(repo_id="mistralai/Mixtral-8x7B-Instruct-v0.1", model_kwargs={"temperature":0.2, "max_length":1000})

    # Similarly, add a test case for "gemini-pro" model
"""

# Add any additional tests as needed
